backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1
pypdf==4.2.0
Pillow==10.3.0
python-docx==1.1.2
openpyxl==3.1.2
python-pptx==0.6.23
img2pdf==0.5.1
reportlab==4.2.0
httpx==0.27.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.26.0
pdf2docx==0.5.8
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
numpy==1.26.4
aiohttp==3.9.5
celery==5.3.6
redis==5.0.4